+++ cmd/server/main.go
@@ -16,7 +16,7 @@ func main() {
 	}
 	done := make(chan os.Signal, 1)
 	signal.Notify(done, os.Interrupt, syscall.SIGINT, syscall.SIGTERM)
-	go storage.InitSaveMem(server.FileStoragePath, server.Restore, server.StoreInterval, done)
+	go storage.InitFileSave(server.FileStoragePath, server.Restore, server.StoreInterval, done)
 	r := handlers.Routers()
 	err = r.Run(server.FlagRunAddr)
 	if err != nil {
