+++ internal/agent/server/sendMetrics.go
@@ -8,7 +8,6 @@ import (
 )
 
 func SendMetrics(memStorage *storage.MemStorage, servAddr string, logger *zap.Logger) {
-	metrics := make([]models.Metrics, 0)
 	for _, store := range memStorage.Metrics {
 		for k, v := range store.Gauge {
 			model := models.Metrics{
@@ -16,21 +15,26 @@ func SendMetrics(memStorage *storage.MemStorage, servAddr string, logger *zap.Lo
 				MType: "gauge",
 				Value: &v,
 			}
-			metrics = append(metrics, model)
+			url := fmt.Sprintf("http://%s/update/", servAddr)
+			req := NewServer(url)
+			code, err := req.SendReqPost("POST", model, logger)
+			if code != 200 || err != nil {
+				logger.Error("Failed to send metrics", zap.Int("code", code),
+					zap.String("url", url), zap.Error(err))
+			}
 		}
 		model := models.Metrics{
 			ID:    "PollCount",
 			MType: "counter",
 			Delta: &store.Counter,
 		}
-		metrics = append(metrics, model)
-	}
-	url := fmt.Sprintf("http://%s/update/", servAddr)
-	req := NewServer(url)
-	code, err := req.SendReqPost("POST", metrics, logger)
-	if code != 200 || err != nil {
-		logger.Error("Failed to send metrics", zap.Int("code", code),
-			zap.String("url", url), zap.Error(err))
+		url := fmt.Sprintf("http://%s/update/", servAddr)
+		req := NewServer(url)
+		code, err := req.SendReqPost("POST", model, logger)
+		if code != 200 || err != nil {
+			logger.Error("Failed to send metrics", zap.Int("code", code),
+				zap.String("url", url), zap.Error(err))
+		}
 	}
 	memStorage.ClearMemStorage()
 }
