+++ internal/server/flags.go
@@ -10,14 +10,12 @@ var FlagRunAddr string
 var StoreInterval int
 var FileStoragePath string
 var Restore bool
-var DBURL string
 
 func ParsFlags() error {
 	flag.StringVar(&FlagRunAddr, "a", "127.0.0.1:8080", "address and port to run server")
 	flag.IntVar(&StoreInterval, "i", 300, "interval before save settings")
 	flag.StringVar(&FileStoragePath, "f", "/tmp/metrics-db.json", "location of metrics db file")
 	flag.BoolVar(&Restore, "r", true, "restore metrics db")
-	flag.StringVar(&DBURL, "d", "", "database url")
 	flag.Parse()
 	var err error
 	if envRunAddr := os.Getenv("ADDRESS"); envRunAddr != "" {
@@ -37,9 +35,6 @@ func ParsFlags() error {
 		if err != nil {
 			return err
 		}
-	}
-	if envDBURL := os.Getenv("DATABASE_DSN"); envDBURL != "" {
-		DBURL = envDBURL
 	}
 	return nil
 }
