+++ internal/server/handlers/routers.go
@@ -19,11 +19,9 @@ func Routers() *gin.Engine {
 	r.Use(middleware.Gzip())
 	r.POST("/update/:metricType/:metricName/:metricValue", UpdateMetricsParam)
 	r.POST("/update/", UpdateMetrics)
-	r.POST("/updates/", UpdateMetrics)
 	r.GET("/", AllRecordsHandler)
 	r.POST("/value/", GetPostMetric)
 	r.GET("/value/:metricType/:metricName", GetMetric)
-	r.GET("/ping", CheckDBConnection)
 
 	return r
 }
