+++ internal/server/models/requestMetrics.go
@@ -6,7 +6,3 @@ type Metrics struct {
 	Delta *int64   `json:"delta,omitempty"` // значение метрики в случае передачи counter
 	Value *float64 `json:"value,omitempty"` // значение метрики в случае передачи gauge
 }
-
-type BatchMetrics struct {
-	Metrics []Metrics `json:"metrics"`
-}
