+++ internal/server/service/getMetric_test.go
@@ -171,9 +171,7 @@ func TestUpdateMetrics(t *testing.T) {
 	storage.ClearStorage()
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			batch := make([]models.Metrics, 1)
-			batch[0] = tt.args
-			_, err := UpdateMetrics(batch)
+			_, err := UpdateMetrics(tt.args)
 			if err != nil {
 				assert.EqualError(t, err, tt.wantErr.Error())
 			}
