+++ internal/server/service/updateMetrics.go
@@ -5,39 +5,35 @@ import (
 	"github.com/stranik28/MetricsCollector/internal/server/storage"
 )
 
-func UpdateMetrics(reqModels []models.Metrics) ([]models.Metrics, error) {
-	respModels := make([]models.Metrics, len(reqModels))
-	for _, reqModel := range reqModels {
-		switch reqModel.MType {
-		case "counter":
-			val, ok := storage.GetMemStorage(reqModel.ID)
-			if !ok {
-				val = storage.Metric{
-					Gauge:   0,
-					Counter: *reqModel.Delta,
-				}
-			} else {
-				val.Counter += *reqModel.Delta
+func UpdateMetrics(reqModel models.Metrics) (models.Metrics, error) {
+	switch reqModel.MType {
+	case "counter":
+		val, ok := storage.GetMemStorage(reqModel.ID)
+		if !ok {
+			val = storage.Metric{
+				Gauge:   0,
+				Counter: *reqModel.Delta,
 			}
-			storage.SetMemStorage(reqModel.ID, val)
-			reqModel.Value = &val.Gauge
-		case "gauge":
-			val, ok := storage.GetMemStorage(reqModel.ID)
-			if !ok {
-				val = storage.Metric{
-					Gauge:   *reqModel.Value,
-					Counter: 0,
-				}
-			} else {
-				val.Gauge = *reqModel.Value
-			}
-			storage.SetMemStorage(reqModel.ID, val)
-			reqModel.Value = &val.Gauge
-		default:
-			err := storage.ErrorIncorrectTypeMetrics
-			return respModels, err
+		} else {
+			val.Counter += *reqModel.Delta
 		}
-		respModels = append(respModels, reqModel)
+		storage.SetMemStorage(reqModel.ID, val)
+		reqModel.Value = &val.Gauge
+	case "gauge":
+		val, ok := storage.GetMemStorage(reqModel.ID)
+		if !ok {
+			val = storage.Metric{
+				Gauge:   *reqModel.Value,
+				Counter: 0,
+			}
+		} else {
+			val.Gauge = *reqModel.Value
+		}
+		storage.SetMemStorage(reqModel.ID, val)
+		reqModel.Value = &val.Gauge
+	default:
+		err := storage.ErrorIncorrectTypeMetrics
+		return reqModel, err
 	}
-	return respModels, nil
+	return reqModel, nil
 }
